requirements.txt
tabula-py
jpype1
dotenv
psycopg2-binary
tabula